rename placeGuideWithCreatorPair to PlaceGuideInfo
